draft for collection
I am just collecting the open events with a filename. We do not have
closes so we cannot time things, but that is not the goal here. I
am also limiting to a container cgroup id instead of trying
to do a bunch of parsing of processes and getting confused with tid
tgid pid ppid and then the fact that the kernel pid != the pid we
understand.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>